+++ src/components/icons/GlobeIcon.jsx
